requirements.txt
django==1.9
psycopg2
djangorestframework==3.4.4
django-rest-swagger
